Handle tokens correctly in the new zone change code.
